+++ src/xyz/oihana/schema/constants/traits/products/PricingCondition.php
@@ -13,13 +13,19 @@ trait PricingCondition
 {
     public const string ADDITIONAL_PROPERTY = 'additionalProperty' ;
     public const string ADJUSTMENT          = 'adjustment' ;
+    public const string CATEGORY            = 'category' ;
+    public const string CUSTOMER            = 'customer' ;
     public const string EXCLUDED_CUSTOMERS  = 'excludedCustomers' ;
     public const string EXCLUDED_PRODUCTS   = 'excludedProducts' ;
     public const string FIXED_PRICE         = 'fixedPrice' ;
     public const string FREE                = 'free' ;
+    public const string PRODUCT             = 'product' ;
+    public const string PROVIDER            = 'provider' ;
     public const string QUANTITY_DISCOUNT   = 'quantityDiscount' ;
     public const string SELECTOR            = 'selector' ;
+    public const string SUBSIDIARY          = 'subsidiary' ;
     public const string SUBSTITUTES_SEGMENT = 'substitutesSegment' ;
     public const string VALID_FROM          = 'validFrom' ;
     public const string VALID_THROUGH       = 'validThrough' ;
+    public const string WAREHOUSE           = 'warehouse' ;
 }